도메인 클래스 엔티티 domain 모듈에 이전
도메인 모듈 빌드 성공 (bootJar X, jar O)
