+++ domain/src/main/java/org/closet/domain/base/BaseUpdatableEntity.java
@@ -1,0 +1,19 @@
+package org.closet.domain.base;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import java.time.Instant;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.LastModifiedDate;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@MappedSuperclass
+public abstract class BaseUpdatableEntity extends BaseEntity {
+
+    @LastModifiedDate
+    @Column(columnDefinition = "timestamp with time zone")
+    private Instant updatedAt;
+}
