+++ domain/src/main/java/org/closet/domain/clothes/ClothesType.java
@@ -1,0 +1,16 @@
+package org.closet.domain.clothes;
+
+public enum ClothesType {
+    TOP,
+    BOTTOM,
+    DRESS,
+    OUTER,
+    UNDERWEAR,
+    ACCESSORY,
+    SHOES,
+    SOCKS,
+    HAT,
+    BAG,
+    SCARF,
+    ETC
+}
