+++ domain/src/main/java/org/closet/domain/feed/Feed.java
@@ -1,0 +1,74 @@
+package org.closet.domain.feed;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.closet.domain.base.BaseUpdatableEntity;
+import org.closet.domain.clothes.Clothes;
+import org.closet.domain.user.User;
+import org.closet.domain.weather.Weather;
+import org.hibernate.annotations.BatchSize;
+
+@Getter
+@Entity
+@Table(name = "feeds")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Feed extends BaseUpdatableEntity {
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "author_id")
+    private User author;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "weather_id", nullable = false)
+    private Weather weather;
+
+    @Column(name = "content", columnDefinition = "TEXT", nullable = false)
+    private String content;
+
+    @BatchSize(size = 100)
+    @OneToMany(mappedBy = "feed", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<FeedClothes> feedClothesList = new ArrayList<>();
+
+    @Column(name = "like_count", nullable = false)
+    private int likeCount = 0;
+
+    public Feed(User author, Weather weather, String content) {
+        this.author = author;
+        this.weather = weather;
+        this.content = content;
+    }
+
+    public void addClothes(Clothes clothes) {
+        FeedClothes feedClothes = new FeedClothes(this, clothes);
+        this.feedClothesList.add(feedClothes);
+    }
+
+    public void updateContent(String content) {
+        if (content == null || content.isBlank()) {
+            throw new IllegalArgumentException("Content cannot be null or blank");
+        }
+        this.content = content;
+    }
+
+    public void incrementLikeCount() {
+        this.likeCount++;
+    }
+
+    public void decrementLikeCount() {
+        if (this.likeCount <= 0) {
+            throw new IllegalStateException("Like count cannot be negative");
+        }
+        this.likeCount--;
+    }
+}
