+++ domain/src/main/java/org/closet/domain/feed/FeedClothes.java
@@ -1,0 +1,39 @@
+package org.closet.domain.feed;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.util.UUID;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.closet.domain.clothes.Clothes;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "feed_clothes")
+public class FeedClothes {
+
+    @Id
+    @GeneratedValue
+    private UUID id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "feed_id", nullable = false)
+    private Feed feed;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "clothes_id", nullable = false)
+    private Clothes clothes;
+
+    public FeedClothes(Feed feed, Clothes clothes) {
+        this.feed = feed;
+        this.clothes = clothes;
+    }
+
+}
