+++ domain/src/main/java/org/closet/domain/feed/FeedLike.java
@@ -1,0 +1,39 @@
+package org.closet.domain.feed;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.util.UUID;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.closet.domain.user.User;
+
+@Getter
+@Entity
+@Table(name = "feed_likes")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class FeedLike {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "feed_id", nullable = false)
+    private Feed feed;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    public FeedLike(Feed feed, User user) {
+        this.feed = feed;
+        this.user = user;
+    }
+}
