+++ domain/src/main/java/org/closet/domain/notification/NotificationLevel.java
@@ -1,0 +1,7 @@
+package org.closet.domain.notification;
+
+public enum NotificationLevel {
+    INFO,
+    WARNING,
+    ERROR
+}
