+++ domain/src/main/java/org/closet/domain/user/Gender.java
@@ -1,0 +1,7 @@
+package org.closet.domain.user;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    OTHER
+}
