+++ domain/src/main/java/org/closet/domain/user/Role.java
@@ -1,0 +1,6 @@
+package org.closet.domain.user;
+
+public enum Role {
+    ADMIN,
+    USER
+}
