+++ domain/src/main/java/org/closet/domain/weather/AsWord.java
@@ -1,0 +1,8 @@
+package org.closet.domain.weather;
+
+public enum AsWord {
+    WEAK,
+    MODERATE,
+    STRONG,
+    UNKNOWN;
+}
