+++ domain/src/main/java/org/closet/domain/weather/PrecipitationType.java
@@ -1,0 +1,11 @@
+package org.closet.domain.weather;
+
+public enum PrecipitationType {
+    // 없음(0), 비(1), 비/눈(2), 눈(3), 소나기(4)
+    NONE,
+    RAIN,
+    RAIN_SNOW,
+    SNOW,
+    SHOWER,
+    UNKNOWN;
+}
