+++ domain/src/main/java/org/closet/domain/weather/SkyStatus.java
@@ -1,0 +1,9 @@
+package org.closet.domain.weather;
+
+public enum SkyStatus {
+    // 맑음, 구름 많음, 흐림
+    CLEAR,
+    MOSTLY_CLOUDY,
+    CLOUDY,
+    UNKNOWN;
+}
