--- src/main/java/project/closet/domain/clothes/entity/ClothesType.java
@@ -1,16 +1,0 @@
-package project.closet.domain.clothes.entity;
-
-public enum ClothesType {
-    TOP,
-    BOTTOM,
-    DRESS,
-    OUTER,
-    UNDERWEAR,
-    ACCESSORY,
-    SHOES,
-    SOCKS,
-    HAT,
-    BAG,
-    SCARF,
-    ETC
-}
