--- src/main/java/project/closet/feed/entity/FeedClothes.java
@@ -1,39 +1,0 @@
-package project.closet.feed.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import project.closet.domain.clothes.entity.Clothes;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "feed_clothes")
-public class FeedClothes {
-
-    @Id
-    @GeneratedValue
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "feed_id", nullable = false)
-    private Feed feed;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "clothes_id", nullable = false)
-    private Clothes clothes;
-
-    public FeedClothes(Feed feed, Clothes clothes) {
-        this.feed = feed;
-        this.clothes = clothes;
-    }
-
-}
