--- src/main/java/project/closet/notification/entity/NotificationLevel.java
@@ -1,7 +1,0 @@
-package project.closet.notification.entity;
-
-public enum NotificationLevel {
-    INFO,
-    WARNING,
-    ERROR
-}
