--- src/main/java/project/closet/user/entity/Gender.java
@@ -1,7 +1,0 @@
-package project.closet.user.entity;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
