--- src/main/java/project/closet/user/entity/Role.java
@@ -1,6 +1,0 @@
-package project.closet.user.entity;
-
-public enum Role {
-    ADMIN,
-    USER
-}
