--- src/main/java/project/closet/weather/entity/AsWord.java
@@ -1,8 +1,0 @@
-package project.closet.weather.entity;
-
-public enum AsWord {
-    WEAK,
-    MODERATE,
-    STRONG,
-    UNKNOWN;
-}
