--- src/main/java/project/closet/weather/entity/PrecipitationType.java
@@ -1,11 +1,0 @@
-package project.closet.weather.entity;
-
-public enum PrecipitationType {
-    // 없음(0), 비(1), 비/눈(2), 눈(3), 소나기(4)
-    NONE,
-    RAIN,
-    RAIN_SNOW,
-    SNOW,
-    SHOWER,
-    UNKNOWN;
-}
