--- src/main/java/project/closet/weather/entity/SkyStatus.java
@@ -1,9 +1,0 @@
-package project.closet.weather.entity;
-
-public enum SkyStatus {
-    // 맑음, 구름 많음, 흐림
-    CLEAR,
-    MOSTLY_CLOUDY,
-    CLOUDY,
-    UNKNOWN;
-}
